pyproject.toml
[project]
name = "Image-Toolkit"
version = "0.1.0"
description = "Image database and edit toolkit"
readme = "README.md"
requires-python = ">=3.11"

# All your runtime deps
dependencies = [
    "accelerate>=1.12.0",
    "celery>=5.6.0",
    "comtypes>=1.4.13",
    "diffusers>=0.35.2",
    "django>=5.2.7",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "dropbox>=12.0.2",
    "google>=3.0.0",
    "google-api-python-client>=2.187.0",
    "google-auth-oauthlib>=1.2.3",
    "imagehash>=4.3.2",
    "jpype1>=1.6.0",
    "moviepy>=1.0.3",
    "msal>=1.34.0",
    "numpy>=2.3.4",
    "opencv-python>=4.11.0.86",
    "peft>=0.18.0",
    "pillow>=12.0.0",
    "pip>=25.3",
    "psycopg2-binary>=2.9.11",
    "pyside6>=6.10.0",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "screeninfo>=0.8.1",
    "selenium>=4.38.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "transformers>=4.57.2",
    "webdriver-manager>=4.0.2",
]

# (Optional) entry-point so you can run `image-toolkit` from terminal
[project.scripts]
# The entry point is now 'app.py' (found in 'backend/src')
image-toolkit = "app:launch_app"

# Corrected 'dev' dependencies
[dependency-groups]
dev = [
    "black>=25.9.0",
    "cryptography>=46.0.3",
    "mypy>=1.18.2",
    "pyinstaller>=6.16.0",
    "pytest>=8.4.2",
    "ruff>=0.14.3",
]

# BUILD CONFIG
[build-system]
requires = ["setuptools>=61", "wheel"]
build-src = "setuptools.build_meta"

# Tell setuptools EXACTLY what to package
[tool.setuptools.packages.find]
# 1. Tell setuptools to look INSIDE 'backend' for packages
where = ["backend"]

# 2. Include non-code files *inside* the packages
[tool.setuptools.package-data]
# "*" = ["*.ui", "*.qrc", "*.png", "*.ico"]

[tool.pyright]
typeCheckingMode = "off"

[tool.pytest.ini_options]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*", "*Test"]
python_functions = ["test_*", "*_test"]
testpaths = "backend/test"
